focus lock / piezo GetTargetPos related tweaks and bugfixes

+++ PYME/Acquire/ui/positionUI.py
@@ -133,7 +133,7 @@ class PositionSliders(wx.Panel):
 #            else:
             
             pos = poss[pName]
-            self.sliders[ind].SetValue(100*pos)
+            self.sliders[ind].SetValue(int(100 * pos))  # wx sliders, at least as of version 4.0.4 don't like float
             self.sliderLabels[ind].SetLabel(u'%s - %2.3f %s' % (pName,pos, unit))
             
             self.sliders[ind].SetMin(100*self.ranges[pName][0])
